+++ changelog.md
@@ -5,6 +5,7 @@ For the latest version of this document, please see [https://github.com/haskell/
 * Support `hashable-1.4.6.0`.
 * Fix an issue where `Hashable Key` wasn't newtype instance over underlying `Text`,
   so with `-ordered-keymap` there were correctness issues.
+* Add instances for `Data.Semigroup.Sum`, `Product`, `Any`, `All`
 
 ### 2.2.2.0